+++ sentry-android-replay/src/main/java/io/sentry/android/replay/viewhierarchy/ViewHierarchyNode.kt
@@ -3,6 +3,7 @@ package io.sentry.android.replay.viewhierarchy
 import android.annotation.TargetApi
 import android.graphics.Rect
 import android.view.View
+import android.view.ViewParent
 import android.widget.ImageView
 import android.widget.TextView
 import io.sentry.SentryOptions
@@ -261,11 +262,30 @@ sealed class ViewHierarchyNode(
                 return true
             }
 
+            if (!this.isMaskContainer(options) &&
+                this.parent != null &&
+                this.parent.isUnmaskContainer(options)
+            ) {
+                return false
+            }
+
             if (this.javaClass.isAssignableFrom(options.experimental.sessionReplay.unmaskViewClasses)) {
                 return false
             }
 
             return this.javaClass.isAssignableFrom(options.experimental.sessionReplay.maskViewClasses)
+        }
+
+        private fun ViewParent.isUnmaskContainer(options: SentryOptions): Boolean {
+            val unmaskContainer =
+                options.experimental.sessionReplay.unmaskViewContainerClass ?: return false
+            return this.javaClass.name == unmaskContainer
+        }
+
+        private fun View.isMaskContainer(options: SentryOptions): Boolean {
+            val maskContainer =
+                options.experimental.sessionReplay.maskViewContainerClass ?: return false
+            return this.javaClass.name == maskContainer
         }
 
         fun fromView(view: View, parent: ViewHierarchyNode?, distance: Int, options: SentryOptions): ViewHierarchyNode {